--- src/main/java/roomescape/controller/GreetingViewController.java
@@ -1,13 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class GreetingViewController {
-
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-}
