--- src/main/java/roomescape/controller/ReservationViewController.java
@@ -1,12 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ReservationViewController {
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "new-reservation";
-    }
-}
