--- src/main/java/roomescape/dao/ReservationDao.java
@@ -1,76 +1,0 @@
-package roomescape.dao;
-
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
-import roomescape.domain.Time;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
-import java.util.Map;
-
-@Repository
-public class ReservationDao {
-
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleJdbcInsert;
-
-    public ReservationDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
-                .withTableName("reservation")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    public List<Reservation> findAll() {
-        String sql = """
-                SELECT
-                    r.id as reservation_id,
-                    r.name,
-                    r.date,
-                    t.id as time_id,
-                    t.time as time_value
-                FROM reservation as r
-                INNER JOIN time as t ON r.time_id = t.id
-                ORDER BY r.id
-                """;
-        return jdbcTemplate.query(sql, reservationRowMapper());
-    }
-
-    public boolean existsByDateAndTimeId(LocalDate date, Long timeId) {
-        String sql = "SELECT COUNT(1) FROM reservation WHERE date = ? AND time_id = ?";
-        Integer count = jdbcTemplate.queryForObject(sql, Integer.class, date, timeId);
-        return count != null && count > 0;
-    }
-
-    public Reservation save(Reservation reservation) {
-        Number key = simpleJdbcInsert.executeAndReturnKey(Map.of(
-                "name", reservation.getName(),
-                "date", reservation.getDate(),
-                "time_id", reservation.getTime().getId()
-        ));
-        return new Reservation(key.longValue(), reservation.getName(), reservation.getDate(), reservation.getTime());
-    }
-
-    public boolean deleteById(Long id) {
-        String sql = "DELETE FROM reservation WHERE id = ?";
-        int updatedRows = jdbcTemplate.update(sql, id);
-        return updatedRows > 0;
-    }
-
-    private RowMapper<Reservation> reservationRowMapper() {
-        return (resultSet, rowNum) -> new Reservation(
-                resultSet.getLong("reservation_id"),
-                resultSet.getString("name"),
-                resultSet.getObject("date", LocalDate.class),
-                new Time(
-                        resultSet.getLong("time_id"),
-                        resultSet.getObject("time_value", LocalTime.class)
-                )
-        );
-    }
-}
