--- src/main/java/roomescape/dto/ErrorResponse.java
@@ -1,6 +1,0 @@
-package roomescape.dto;
-
-public record ErrorResponse(
-        String message
-) {
-}
