--- src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,20 +1,0 @@
-package roomescape.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(
-        @NotBlank(message = "이름은 필수 입력값입니다.")
-        String name,
-
-        @NotNull(message = "날짜는 필수 입력값입니다.")
-        @JsonFormat(pattern = "yyyy-MM-dd")
-        LocalDate date,
-
-        @NotNull(message = "시간은 필수 입력값입니다.")
-        Long timeId
-) {
-}
