--- src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,28 +1,0 @@
-package roomescape.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import roomescape.domain.Reservation;
-
-import java.time.LocalDate;
-
-@Getter
-@AllArgsConstructor
-public class ReservationResponse {
-
-    private Long id;
-    private String name;
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    private LocalDate date;
-    private TimeResponse time;
-
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate(),
-                TimeResponse.from(reservation.getTime())
-        );
-    }
-}
