+++ src/test/java/roomescape/MissionStepTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationController;
+import roomescape.reservation.controller.ReservationController;
 
 import java.lang.reflect.Field;
 import java.util.HashMap;
